+++ src/main/java/com/silverminer/shrines/generators/DataProviderEvents.java
@@ -6,16 +6,10 @@
  * file, You can obtain one at https://mozilla.org/MPL/2.0/.
  */
 
-package com.silverminer.shrines.data.generators;
+package com.silverminer.shrines.generators;
 
 import com.silverminer.shrines.ShrinesMod;
-import com.silverminer.shrines.data.ConfiguredStructureFeatures;
-import com.silverminer.shrines.data.StructureSets;
-import com.silverminer.shrines.data.TemplatePools;
 import net.minecraft.data.info.WorldgenRegistryDumpReport;
-import net.minecraft.world.level.levelgen.feature.StructureFeature;
-import net.minecraftforge.event.RegistryEvent;
-import net.minecraftforge.eventbus.api.EventPriority;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.forge.event.lifecycle.GatherDataEvent;
@@ -26,12 +20,5 @@ public class DataProviderEvents {
    public static void onGatherDataEvent(GatherDataEvent event) {
       event.getGenerator().addProvider(new ShrinesBiomeTagsProvider(event.getGenerator(), event.getExistingFileHelper()));
       event.getGenerator().addProvider(new WorldgenRegistryDumpReport(event.getGenerator()));
-   }
-
-   @SubscribeEvent(priority = EventPriority.LOW)
-   public static void onRegisterBlocks(RegistryEvent.Register<StructureFeature<?>> event) {
-      TemplatePools.bootstrap();
-      ConfiguredStructureFeatures.bootstrap();
-      StructureSets.bootstrap();
    }
 }