+++ changelog.md
@@ -1,3 +1,20 @@
+### 4.0.0-alpha3
+
+**It's not possible to update old worlds and installations to this version**
+**This is an alpha release; breaking changes can happen at any time. Don't use this file in production**
+
+- required Forge 40.0.36+
+- Add lady_jessa's luxury villa model of the modern villa
+- Implement (missed) Ground level delta spawn criteria
+- Add default spawn criteria to all structures (ground level delta)
+- Change harbour's biomes to oceans again
+- Renamed watch tower's and oriental sanctuary's template
+- Add missing templates to nether shrine's template pool
+- Add missing biome tag for world tree manor
+
+**Feedback is appreciated**
+
+===========
 ### 4.0.0-alpha2
 
 **It's not possible to update old worlds and installations to this version**
@@ -14,7 +31,7 @@
 
 **Feedback is appreciated**
 
-===========~~~~
+===========
 ### 4.0.0-alpha1
 
 **It's not possible to update old worlds and installations to this version**